+++ packages/@tinacms/mdx/src/next/tests/mdx-blocks-rich-text-children-with-an-empty-object-value/index.test.ts
@@ -1,0 +1,29 @@
+import { it, expect } from 'vitest'
+import { stringifyMDX } from '../../../stringify'
+import { field } from './field'
+import * as util from '../util'
+import { RootElement } from '../../../parse/plate'
+
+it('matches input', () => {
+  const tree: RootElement = {
+    type: 'root',
+    children: [
+      {
+        type: 'mdxJsxFlowElement',
+        name: 'Cta',
+        children: [
+          {
+            type: 'text',
+            text: '',
+          },
+        ],
+        props: {
+          children: {},
+        },
+      },
+    ],
+  }
+
+  const string = stringifyMDX(tree, field, (v) => v)
+  expect(string).toMatchFile(util.mdPath(__dirname))
+})