+++ packages/tinacms/src/index.ts
@@ -10,6 +10,7 @@ export { useDocumentCreatorPlugin } from './hooks/use-content-creator'
 export { TinaAdmin } from './admin'
 export { RouteMappingPlugin } from './admin/plugins/route-mapping'
 export { TinaAdminApi } from './admin/api'
+export { ErrorDialog } from './admin/components/ErrorDialog'
 
 export * from './toolkit'
 export { Form } from './toolkit/forms/form'