+++ experimental-examples/tina-cloud-starter/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tinacms/starter
 
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies [1563ce5b2]
+- Updated dependencies [e83ba8855]
+  - tinacms@1.5.19
+  - @tinacms/vercel-previews@0.0.15
+
 ## 0.1.14
 
 ### Patch Changes